# apps/luxledger-api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LuxLedger API
  version: 0.2.0
  description: Reference REST API built on top of @lux/ledger.
servers:
  - url: http://localhost:3000
security:
  - BearerAuth: []
paths:
...
  /v1/accounts:
    post:
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    get:
      summary: List accounts
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/LedgerIdQuery'
      responses:
        '200':
          description: Account page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsPage'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /v1/accounts/{id}:
    get:
      summary: Get account by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /v1/accounts/{id}/balance-as-of:
    get:
      summary: Get account balance as of a timestamp
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: at
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Historical account balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceAsOfResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /v1/accounts/{id}/balance-history:
    get:
      summary: List account balance history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Account balance history page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceHistoryResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /v1/holds:
    post:
      summary: Create a funds hold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHoldRequest'
      responses:
        '201':
          description: Hold created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateHoldResponse'
        '200':
          description: Idempotent retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateHoldResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /v1/holds/{id}/commit:
    post:
      summary: Commit all or part of a hold
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitHoldRequest'
      responses:
        '201':
          description: Hold commit created a transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitHoldResponse'
        '200':
          description: Idempotent retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitHoldResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /v1/holds/{id}/void:
    post:
      summary: Void the remaining amount of a hold
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Hold voided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoidHoldResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      required: false
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        minLength: 1
      required: false
    LedgerIdQuery:
      name: ledger_id
      in: query
      schema:
        type: string
        format: uuid
      required: false
...
  responses:
    BadRequestError:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ConflictError:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    TransactionEntryRequest:
      type: object
      additionalProperties: false
      required: [account_id, direction, amount_minor, currency]
      properties:
        account_id:
          type: string
          format: uuid
        direction:
          type: string
          enum: [DEBIT, CREDIT]
        amount_minor:
          type: string
          pattern: '^[1-9][0-9]*$'
        currency:
          type: string
          pattern: '^(?=.*\S).+$'
...
    CreateAccountRequest:
      type: object
      additionalProperties: false
      required: [ledger_id, name, side, currency]
      properties:
        ledger_id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        side:
          type: string
          enum: [DEBIT, CREDIT]
        overdraft_policy:
          type: string
          enum: [ALLOW, DISALLOW]
        currency:
          type: string
          pattern: '^(?=.*\S).+$'
      example:
        ledger_id: 00000000-0000-4000-8000-000000000001
        name: Cash
        side: DEBIT
        overdraft_policy: ALLOW
        currency: USD
    Account:
      type: object
      additionalProperties: false
      required: [id, tenant_id, ledger_id, name, side, overdraft_policy, currency, balance_minor, created_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        ledger_id:
          type: string
          format: uuid
        name:
          type: string
        side:
          type: string
          enum: [DEBIT, CREDIT]
        overdraft_policy:
          type: string
          enum: [ALLOW, DISALLOW]
        currency:
          type: string
        balance_minor:
          type: string
        created_at:
          type: string
          format: date-time
      example:
        id: 00000000-0000-4000-8000-000000000101
        tenant_id: 11111111-1111-4111-8111-111111111111
        ledger_id: 00000000-0000-4000-8000-000000000001
        name: Cash
        side: DEBIT
        overdraft_policy: ALLOW
        currency: USD
        balance_minor: '100'
        created_at: '2026-01-01T00:00:00.000Z'
    AccountsPage:
      type: object
      additionalProperties: false
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        next_cursor:
          type: string
          nullable: true
    BalanceAsOfResponse:
      type: object
      additionalProperties: false
      required: [account_id, timestamp, posted_minor, inflight_debit_minor, inflight_credit_minor, available_minor]
      properties:
        account_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        posted_minor:
          type: string
        inflight_debit_minor:
          type: string
        inflight_credit_minor:
          type: string
        available_minor:
          type: string
    BalanceSnapshot:
      type: object
      additionalProperties: false
      required: [id, tenant_id, ledger_id, account_id, event_type, source_id, posted_minor, inflight_debit_minor, inflight_credit_minor, effective_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        ledger_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        event_type:
          type: string
          enum: [TX_APPLIED, HOLD_CREATED, HOLD_COMMITTED, HOLD_VOIDED, ADJUSTMENT]
        source_id:
          type: string
          format: uuid
        posted_minor:
          type: string
        inflight_debit_minor:
          type: string
        inflight_credit_minor:
          type: string
        effective_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    BalanceHistoryResponse:
      type: object
      additionalProperties: false
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BalanceSnapshot'
        next_cursor:
          type: string
          nullable: true
...
    CreateHoldRequest:
      type: object
      additionalProperties: false
      required: [ledger_id, reference, currency, entries]
      properties:
        ledger_id:
          type: string
          format: uuid
        reference:
          type: string
          pattern: '^(?=.*\S).+$'
        currency:
          type: string
          pattern: '^(?=.*\S).+$'
        description:
          type: string
          pattern: '^(?=.*\S).+$'
        entries:
          type: array
          minItems: 2
          items:
            $ref: '#/components/schemas/TransactionEntryRequest'
    CreateHoldResponse:
      type: object
      additionalProperties: false
      required: [hold_id, created, state, remaining_amount_minor]
      properties:
        hold_id:
          type: string
          format: uuid
        created:
          type: boolean
        state:
          type: string
          enum: [HELD, APPLIED, VOIDED]
        remaining_amount_minor:
          type: string
          pattern: '^[0-9]+$'
    CommitHoldRequest:
      type: object
      additionalProperties: false
      required: [reference]
      properties:
        reference:
          type: string
          pattern: '^(?=.*\S).+$'
        amount_minor:
          type: string
          pattern: '^[1-9][0-9]*$'
    CommitHoldResponse:
      type: object
      additionalProperties: false
      required: [hold_id, transaction_id, created, state, remaining_amount_minor]
      properties:
        hold_id:
          type: string
          format: uuid
        transaction_id:
          type: string
          format: uuid
        created:
          type: boolean
        state:
          type: string
          enum: [HELD, APPLIED]
        remaining_amount_minor:
          type: string
          pattern: '^[0-9]+$'
    VoidHoldResponse:
      type: object
      additionalProperties: false
      required: [hold_id, state, voided, remaining_amount_minor]
      properties:
        hold_id:
          type: string
          format: uuid
        state:
          type: string
          enum: [VOIDED]
        voided:
          type: boolean
        remaining_amount_minor:
          type: string
          pattern: '^[0-9]+$'
...